exception handling added to controller

+++ web/src/main/java/com/expert_soft/controller/CartController.java
@@ -1,102 +1,72 @@
 package com.expert_soft.controller;
 
 
-import com.expert_soft.helper.JsonHelper;
-import com.expert_soft.helper.impl.JsonHelperImpl;
-import com.expert_soft.model.AjaxResponseCart;
 import com.expert_soft.model.Cart;
-import com.expert_soft.model.Phone;
+import com.expert_soft.model.OrderItem;
 import com.expert_soft.service.CartService;
+import com.expert_soft.validator.group.G_Cart;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.io.IOException;
+import java.util.ArrayList;
 
 import static com.expert_soft.model.ServletConstants.*;
 
 @Controller
-@SessionAttributes ({CART_ATTR, CART_CURRICULUM_ATTR})
+@SessionAttributes ({CART_ATTR})
 public class CartController {
 
     private static final Logger LOGGER = Logger.getLogger(CartController.class);
 
     private CartService cartService;
-    private JsonHelper jsonHelper;
+
+//
+//    @Autowired Validator validator;
+//
+//    //Set a form validator
+//    @InitBinder
+//    protected void initBinder(WebDataBinder binder) {
+//        binder.setValidator(validator);
+//    }
 
     @Autowired
     public void setCartService(CartService cartService) {
         this.cartService = cartService;
     }
-    @Autowired
-    public void setJsonHelper(JsonHelperImpl jsonHelper) {
-        this.jsonHelper = jsonHelper;
-    }
 
     @RequestMapping( value = "/cart", method = RequestMethod.GET)
     public ModelAndView cart(ModelMap model){
-        Object cart = model.get(CART_ATTR);
+        Cart cart = (Cart) model.get(CART_ATTR);
         if (cart == null){
             return new ModelAndView("emptyCart");
         }
-        return new ModelAndView("cart");
-    }
-
-    @RequestMapping(value = "/add_to_cart",
-                    produces = MediaType.APPLICATION_JSON_UTF8_VALUE,
-                    method = RequestMethod.GET)
-    public @ResponseBody String addToCart(ModelMap model,
-                               @RequestParam(PHONE_TO_DELETE) Long phoneId,
-                               @RequestParam(QUANTITY_PARAM) Integer quantity){
-        LOGGER.debug("catch ajax changed: modelMap " + model);
-
-        Cart cart = getCurrentCart(model);
-        Phone phone = cartService.addToCart(cart, phoneId, quantity);
-        model.put(CART_ATTR, cart);
-        com.expert_soft.model.excluded.CartCurriculum curriculum = cartService.buildCurriculum(cart);
-        model.put(CART_CURRICULUM_ATTR, curriculum);
-
-        AjaxResponseCart response = new AjaxResponseCart();
-        response.setResult(curriculum);
-        response.setCode("200");
-        response.setMsg("Model of phone " + phone.getModel() + " successfully added to cart");
-
-        LOGGER.debug("cart changed: " + cart.toString() );
-
-        String result = "";
-        try {
-           result= jsonHelper.write(response);
-        } catch (IOException e) {
-           LOGGER.debug("Fail to write " + response.getResult());
-        }
-        return result;
+        int size = cart.getItemsMap().size();
+        return new ModelAndView("fullCart", CART_ITEMS, new OrderItem[size]);
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.POST)
     public String changeQuantity(ModelMap model,
-                                 @ModelAttribute(CART_ATTR) Cart cart,
+                                 @Validated(G_Cart.Item.class)
+                                 @ModelAttribute(CART_ITEMS) OrderItem[] lightCart,
                                  @RequestParam(PHONE_TO_DELETE) Long[] phoneArray,
                                  BindingResult result){
-
-
-        cartService.deleteFromCart(cart, phoneArray);
-        com.expert_soft.model.excluded.CartCurriculum curriculum = cartService.buildCurriculum(cart);
-        model.addAttribute(CART_CURRICULUM_ATTR, curriculum);
-        return "redirect:cart";
-    }
-
-    private Cart getCurrentCart(ModelMap modelMap){
-        Cart cart = (Cart) modelMap.get("cart");
-        if (cart == null){
-            cart = Cart.EMPTY_CART;
+        if (result.hasErrors()){
+            model.addAttribute(CART_ITEMS, new ArrayList<OrderItem>());
+            return "fullCart";
+        } else {
+            Cart cart = (Cart) model.get(CART_ATTR);
+            for (OrderItem item : lightCart){
+                cartService.changeQuantity(cart, item.getKey(), item.getQuantity());
+            }
+            cartService.deleteFromCart(cart, phoneArray);
+            return "redirect:/order";
         }
-        return cart;
     }
-
 
 }
